requirements.txt
Flask==3.0.3
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.0.1
pystray==0.19.5